+++ backend/src/main/java/com/woowacourse/ternoko/common/exception/MemberNotFoundException.java
@@ -1,5 +1,6 @@
 package com.woowacourse.ternoko.common.exception;
 
+import com.woowacourse.ternoko.common.exception.type.ExceptionType;
 import org.springframework.http.HttpStatus;
 
 public class MemberNotFoundException extends CommonException {
