+++ backend/src/main/java/com/woowacourse/ternoko/login/aop/LoginMemberVerifier.java
@@ -1,7 +1,7 @@
 package com.woowacourse.ternoko.login.aop;
 
-import static com.woowacourse.ternoko.common.exception.ExceptionType.COACH_NOT_ALLOWED;
-import static com.woowacourse.ternoko.common.exception.ExceptionType.CREW_NOT_ALLOWED;
+import static com.woowacourse.ternoko.common.exception.type.ExceptionType.COACH_NOT_ALLOWED;
+import static com.woowacourse.ternoko.common.exception.type.ExceptionType.CREW_NOT_ALLOWED;
 
 import com.woowacourse.ternoko.domain.member.MemberType;
 import com.woowacourse.ternoko.login.exception.CoachNotAllowedException;
