+++ backend/src/main/java/com/woowacourse/ternoko/login/application/AuthInterceptor.java
@@ -2,7 +2,7 @@ package com.woowacourse.ternoko.login.application;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
-import com.woowacourse.ternoko.common.exception.ExceptionType;
+import com.woowacourse.ternoko.common.exception.type.ExceptionType;
 import com.woowacourse.ternoko.common.exception.InvalidTokenException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
