+++ backend/src/main/java/com/woowacourse/ternoko/login/exception/CrewNotAllowedException.java
@@ -1,7 +1,7 @@
 package com.woowacourse.ternoko.login.exception;
 
 import com.woowacourse.ternoko.common.exception.CommonException;
-import com.woowacourse.ternoko.common.exception.ExceptionType;
+import com.woowacourse.ternoko.common.exception.type.ExceptionType;
 import org.springframework.http.HttpStatus;
 
 public class CrewNotAllowedException extends CommonException {
